Make vector types immutable
This avoids the problem of e.g. character_body.velocity.x += 10 not
updating the character_body.velocity property. It should be possible to
work around the inconvenience of updating vector components by using
macros like the packages Setfield.jl and Accessors.jl do, but with
support for properties in addition to fields.